+++ server/src/completion/basic.ts
@@ -1,5 +1,5 @@
 import { CompletionItem, Position, Range, TextDocumentPositionParams } from "vscode-languageserver";
-import { documents } from "../server";
+import { documents, programs } from "../server";
 import { onCommandCompletion } from "./command";
 
 export function onCompletionImpl(positionParams: TextDocumentPositionParams): CompletionItem[] {
@@ -10,11 +10,12 @@ export function onCompletionImpl(positionParams: TextDocumentPositionParams): Co
       Position.create(positionParams.position.line, 0), 
       Position.create(positionParams.position.line, positionParams.position.character)));
   const splitResult = line.split(/\s+/).length;
+  const program = programs.get(positionParams.textDocument.uri);
   // 1 - label
   // 2 - command
   // 3 - after the command
   if (splitResult == 2)
-    return onCommandCompletion();
+    return onCommandCompletion(program);
   return [];
 }