+++ server/src/completion/command.ts
@@ -1,10 +1,13 @@
 import { CompletionItem, CompletionItemKind } from "vscode-languageserver";
 import { operations } from "../dasm/operations";
 import { NAMES } from "../dasm/directives";
+import { Program } from "../program";
 
-export function onCommandCompletion(): CompletionItem[] {
+export function onCommandCompletion(program: Program | undefined): CompletionItem[] {
     const result = operationCompletion();
     result.push(...directivesCompletion());
+    if (program)
+        result.push(...macrosCompletion(program));
     return result;
 }
 
@@ -12,7 +15,7 @@ function operationCompletion(): CompletionItem[] {
     return Object.entries(operations)
         .map(([key, value]) => { return {
             label: key,
-            kind: CompletionItemKind.Keyword,
+            kind: CompletionItemKind.Operator,
             documentation: value.title
         };});
 }
@@ -22,5 +25,13 @@ function directivesCompletion(): CompletionItem[] {
         .map(it => { return {
             label: it,
             kind: CompletionItemKind.Keyword,
+        };});
+}
+
+function macrosCompletion(program: Program): CompletionItem[] {
+    return Array.from(program.macroses.keys())
+        .map(it => { return {
+            label: it,
+            kind: CompletionItemKind.Method,
         };});
 }